set recovery test working
