+++ package.json
@@ -9,8 +9,8 @@
     "prepublish": "yarn run prepare",
     "test": "yarn run deploy && mocha tests/*.test.ts -r ts-node/register",
     "deploy": "truffle deploy --reset",
-    "testrpc": "testrpc --account='0xa2985156e3da3046101ae2b26093f647fed9d6ce31ffeb4d975d143ac69e5b46,1000000000000000000000001' --account='0x217bd5bc9087a7349c771bc78da459bf2862c804eabbf3b54f16193c67f9a95d,5000000000000000000000001'"
-    },
+    "testrpc": "ganache-cli --account='0xa2985156e3da3046101ae2b26093f647fed9d6ce31ffeb4d975d143ac69e5b46,1000000000000000000000001' --account='0x217bd5bc9087a7349c771bc78da459bf2862c804eabbf3b54f16193c67f9a95d,5000000000000000000000001'"
+  },
   "files": [
     "js/",
     "build/"
@@ -25,7 +25,7 @@
     "@types/node": "^9.4.0",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
-    "ethereumjs-testrpc": "^6.0.3",
+    "ganache-cli": "^6.4.3",
     "mocha": "^5.0.0",
     "truffle": "4.1.14",
     "truffle-hdwallet-provider": "0.0.3",
